Moving App.js to routes folder; new place for cacheImages (HelperFunctions.js)
