+++ frontend/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App';
+import App from './routes/App';
 import AuthContextProvider from './contexts/AuthContext';
 import AppStateContextProvider from './contexts/AppStateContext';
 import 'mapbox-gl/dist/mapbox-gl.css';
